+++ src/baseline.rs
@@ -8,7 +8,6 @@ pub mod dora_compile;
 pub mod dora_entry;
 pub mod dora_native;
 pub mod dora_throw;
-pub mod expr;
 pub mod fct;
 pub mod info;
 pub mod map;
