+++ src/baseline/ast.rs
@@ -1,4 +1,5 @@
 pub use self::codegen::AstCodeGen;
 
 mod codegen;
+mod expr;
 pub mod info;
